NLPBackend/requirements.txt
aiohttp
aiohttp-cors
pytorch
gensim==3.8.0
tensorboardX
cytoolz
py-rouge
